Split compare page metadata
